pyproject.toml
[project]
name = "omi-cli"
version = "0.1.0"
description = "Command-line interface for the Omi AI wearable developer API"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Blake" },
    { name = "omi-cli contributors" },
]
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "rich>=13.7",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
    "keyring>=24.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
omi = "omi_cli.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/omi_cli"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
